Implement ILogger for console logs
In this commit new functionality was added to be able to have logs across the board made by ILogger system. This is to enable ILogger to log on the console of the machine. spdlog has been added as a submodule. Console logs have been introduced across the board with ILogger. Also, as part of cleaning up for introducing ILogger, for instance, all the unnecessary logs have been removed from GltfLoader.cpp. Moreover, as part of integrating ILogger, message args are appended in one single string rather than multiple std out operations.