Add Stats module to UI and optimize StatisticsManager
The modifications to MoonshineApp.cpp add a new dock window labeled "Stats" to the application's UI. The changes in StatisticsManager.cpp replace the use of InputInt and InputDouble functions with Text functions to display statistics, improving readability. The latter also includes the use of the CollapsingHeader functionality to provide a more organized display. The loadSettings function has been removed from the MoonshineApp() constructor for optimization purposes.