Improve glTF model import and add scene save and load functionality
This revision overhauls the glTF model import system to handle filename and path metadata, object renaming, and transformation overwrites. A warning is added for when a glTF file with multiple models is loaded. Furthermore, two new features were implemented: the ability to serialize a scene into a JSON object and save it to a file, and conversely, load a scene from a JSON file. This allows for saving and loading of workspaces. Additional tweaks include a minor code reformat and inclusion of new utility functions.