Introduce the easy profiler for performance tuning.
The changes include adding easy_profiler to the project as a git submodule, linking it in CMakeLists.txt, and integrating it into MoonshineApp.cpp to profile different sections of the main loop. Custom profiling blocks were created to target areas of interest, and profiling data is dumped to a file.